+++ src/pkgmt/deprecation.py
@@ -5,16 +5,16 @@ from collections import defaultdict
 
 import click
 
-from pkgmt.versioner.versionersetup import VersionerSetup
 from pkgmt.versioner.util import complete_version_string
 from pkgmt.exceptions import ProjectValidationError
+from pkgmt.versioner.versioner import Versioner
 
 
 class Deprecations:
     def __init__(self, root_dir=None) -> None:
         self.root_dir = root_dir
 
-        versioner = VersionerSetup(project_root=root_dir)
+        versioner = Versioner(project_root=root_dir)
         self.current = complete_version_string(
             versioner.current_version().replace("dev", "")
         )