+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * [Feature] Added file path to `pkgmt lint` ([#56](https://github.com/ploomber/pkgmt/issues/56))
 * [Feature] Clearer error when missing `pyproject.toml`
 * [Feature] Added `--exclude` option to pkgmt `lint` and pkgmt `format` ([#55](https://github.com/ploomber/pkgmt/issues/55))
+* [API Change] Support for projects containing `version` key in `pyproject.toml`. (#58)
+
 
 ## 0.6.1 (2023-06-09)